:truck: Move signature parsing to caikit.core
