+++ caikit/runtime/service_generation/create_service.py
@@ -28,7 +28,7 @@ from ... import get_config
 from .core_module_helpers import get_module_info
 from .primitives import is_primitive_method
 from .rpcs import CaikitRPCBase, ModuleClassTrainRPC, TaskPredictRPC
-from .signature_parsing.module_signature import CaikitCoreModuleMethodSignature
+from caikit.core.signature_parsing.module_signature import CaikitCoreModuleMethodSignature
 from caikit.core.module import ModuleBase
 
 log = alog.use_channel("CREATE-RPCS")
