+++ tests/core/signature_parsing/test_parsers.py
@@ -22,11 +22,9 @@ import inspect
 
 # Third Party
 from docstring_parser import ParseError
-import pytest
 
 # Local
-from caikit.runtime.service_generation.signature_parsing import docstrings
-from caikit.runtime.service_generation.signature_parsing.parsers import (
+from caikit.core.signature_parsing.parsers import (
     _get_dm_type_from_name,
     _snake_to_camel,
     get_args_with_defaults,
